dataflow_flex_template_file_transform/requirements.txt
apache-beam[gcp]
google-auth
requests